requirements.txt
aiogram==3.4.1
google-generativeai==0.8.3
python-dotenv==1.0.1
numpy>=2.0.0
pandas>=2.2.0
openpyxl==3.1.2
opencv-python>=4.9.0.80
Pillow>=10.2.0
pdf2image==1.17.0
aiofiles==23.2.1